move to legacy: server_error_handling
