+++ weave/weave_server.py
@@ -44,7 +44,7 @@ from weave.legacy import (
     weavejs_fixes,
 )
 from weave.legacy.language_features.tagging import tag_store
-from weave.server_error_handling import client_safe_http_exceptions_as_werkzeug
+from weave.legacy.server_error_handling import client_safe_http_exceptions_as_werkzeug
 
 logger = logging.getLogger(__name__)
